+++ packages/blockchain/src/blockchain.ts
@@ -22,8 +22,7 @@ import {
 import { DBManager } from './db/manager.js'
 import { DBTarget } from './db/operation.js'
 
-import type { Consensus } from './consensus/index.js'
-import type { BlockchainInterface, BlockchainOptions, OnBlock } from './types.js'
+import type { BlockchainInterface, BlockchainOptions, Consensus, OnBlock } from './types.js'
 import type { BlockData } from '@ethereumjs/block'
 import type { CliqueConfig } from '@ethereumjs/common'
 import type { BigIntLike, DB, DBObject, GenesisState } from '@ethereumjs/util'
