+++ packages/blockchain/src/consensus/ethash.ts
@@ -2,7 +2,7 @@ import { ConsensusAlgorithm } from '@ethereumjs/common'
 import { Ethash } from '@ethereumjs/ethash'
 
 import type { Blockchain } from '../index.js'
-import type { Consensus, ConsensusOptions } from './interface.js'
+import type { Consensus, ConsensusOptions } from '../types.js'
 import type { Block, BlockHeader } from '@ethereumjs/block'
 
 /**
