+++ packages/blockchain/src/consensus/index.ts
@@ -1,6 +1,5 @@
 import { CasperConsensus } from './casper.js'
 import { CliqueConsensus } from './clique.js'
 import { EthashConsensus } from './ethash.js'
-import { Consensus } from './interface.js'
 
-export { CasperConsensus, CliqueConsensus, Consensus, EthashConsensus }
+export { CasperConsensus, CliqueConsensus, EthashConsensus }
